+++ app/src/main/java/com/valdesekamdem/taskflow/feature/task/viewmodel/EditTaskViewModel.kt
@@ -105,7 +105,7 @@ class EditTaskViewModel @AssistedInject constructor(
         viewModelScope.launch {
             try {
                 if (screen.id != null) {
-                    taskRepository.updateTask(screen.id.toLong(), taskModel)
+                    taskRepository.updateTask(screen.id, taskModel)
                 } else {
                     taskRepository.addTask(taskModel)
                 }
